+++ CHANGELOG.md
@@ -3,6 +3,9 @@ Change Log
 
 0.5.0 (in development) 
 ----------------------
+- 2020-05-11: Add outputs to DeGrooteFregly2016Muscle for passive elastic and 
+              damping forces.
+
 - 2020-05-11: Fix an updated function name in exampleSquatToStand.m.
 
 - 2020-04-16: Added exampleKinematicConstraints.py to visualize how Moco handles